Refactored pooling structure
