+++ Assets/Fiber/Packages/Signals/Runtime/DynamicDependencies.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2e4894ce7015f49c0a5be5ef2aa82c19
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
